cleanup: remove no-op error_handling extractor shells + dead try regexes
Closes #2331
Closes #2332

After PR #2323 dropped try_catch pattern emission, several remnants
remained: no-op extractErrorHandlingPatterns functions in py/java/js
extractor.go call sites, and the ehTryStatementRE / ehPyTryStatementRE
regexes wired into AppliesTo() in patterns/error_handling_detector.go.

This deletes all of them. Other error_handling sub-patterns
(go_error_return, panic_recover, rust_match_result, rust_question_mark,
elixir_ok_error) are unaffected.
